requirements.txt
discord.py
Pillow
numpy
pandas
lxml
requests
bs4
pyshark
nest_asyncio
pyautogui